conf: set LAYERSERIES_COMPAT to thud
The LAYERSERIES_COMPAT variable is meant to indicate the list of
compatible version(s) of OE-core. This branch of meta-qti-core is
compatible with thud only.

OE-core will check the value of LAYERSERIES_COMPATH with
LAYERSERIES_COMPAT_core, and warn in case an incompatbility is
detected.

This change fixes the following warning:
WARNING: Layer msm-core should set LAYERSERIES_COMPAT_msm-core in its
conf/layer.conf file to list the core layer names it is compatible
with.

It was tested against LE.UM.3.4.1.

Change-Id: I955c672aef58cc8f6cbc7a40500bc1a60d3c1b9c
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>